cloud/cloud_sql/app/requirements.txt
fastapi[all]
asyncpg
sqlalchemy[asyncio]
python-dotenv
